published version

+++ Kiosk_App/Kiosk_App/src/de/nexxoo/kiosk_app/HistoryActivity.java
@@ -1,7 +1,9 @@
 package de.nexxoo.kiosk_app;
 
 import android.app.Activity;
+import android.app.AlertDialog;
 import android.content.Context;
+import android.content.DialogInterface;
 import android.content.Intent;
 import android.graphics.Color;
 import android.graphics.drawable.ColorDrawable;
@@ -26,6 +28,7 @@ import de.nexxoo.kiosk_app.entity.Video;
 import de.nexxoo.kiosk_app.layout.*;
 import de.nexxoo.kiosk_app.tools.FileStorageHelper;
 import de.nexxoo.kiosk_app.tools.Global;
+import de.nexxoo.kiosk_app.tools.Misc;
 import de.nexxoo.kiosk_app.tools.Nexxoo;
 import de.nexxoo.kiosk_app.webservice.NexxooWebservice;
 import de.nexxoo.kiosk_app.webservice.OnJSONResponse;
@@ -34,6 +37,7 @@ import org.json.JSONObject;
 
 import java.io.File;
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 
 /**
@@ -52,6 +56,8 @@ public class HistoryActivity extends Activity {
 
 	private boolean isVideoDownloaded;
 	public static String CONTENTTYPE = "contentTypeId";
+	private boolean isWifiOn;
+	private boolean fromMenu;
 
 
 	@Override
@@ -64,7 +70,9 @@ public class HistoryActivity extends Activity {
 		super.onCreate(savedInstanceState);
 		setContentView(R.layout.history);
 		mContext = this;
-
+		Intent i = getIntent();
+		isWifiOn = i.getBooleanExtra(getString(R.string.wifi_status), true);
+		fromMenu = i.getBooleanExtra(getString(R.string.from_menu_click),false);
 
 		mFileStorgeHelper = new FileStorageHelper(mContext);
 		listview = (SwipeMenuListView) findViewById(R.id.history_list);
@@ -78,16 +86,19 @@ public class HistoryActivity extends Activity {
 		((ViewGroup) listview.getParent()).addView(emptyView);
 		listview.setEmptyView(emptyView);
 
-		getHistoryContentsFromWebServer();
+		if (isWifiOn)
+			getHistoryContentsFromWebServer();
+		else
+			getHistoryContentsFromDB();
 		initSwipeListView(Global.isNormalScreenSize);
 	}
 
-	private Integer[] getContentIdFromDB() {
+	private Integer[] getContentIdFromSharedPreferences() {
 		return Nexxoo.getContentIds(mContext);
 	}
 
 	private void getHistoryContentsFromWebServer() {
-		Integer[] ids = getContentIdFromDB();
+		Integer[] ids = getContentIdFromSharedPreferences();
 
 		NexxooWebservice.getContentByIds(true, ids, new OnJSONResponse() {
 			@Override
@@ -112,10 +123,52 @@ public class HistoryActivity extends Activity {
 	private void getHistoryContentsFromDB() {
 		ContentDBHelper db = new ContentDBHelper(this);
 		List<Content> list = db.getAllContacts();
+		convertContentToBaseEntity(list);
+		listAdapter = new HistoryListAdapter(mContext, Global.isNormalScreenSize ? R.layout
+				.history_listview_item : R.layout.history_listview_item_big, mBaseEntityList, isWifiOn);
+		listview.setAdapter(listAdapter);
 	}
 
-	private void convertContentToBaseEntity(List<Content> list){
-		for(Content content :list){
+	private void convertContentToBaseEntity(List<Content> list) {
+		mBaseEntityList.clear();
+		Collections.reverse(list);
+		for (Content content : list) {
+			int contentTypeId = content.getContentTypeId();
+			switch (contentTypeId) {
+				case 1:
+					Catalog catalog = new Catalog();
+					catalog.setContentId(content.getContentId());
+					catalog.setName(content.getName());
+					catalog.setUrl(content.getUrl());
+					catalog.setFileName(content.getFileName());
+					catalog.setSize(content.getSize());
+					catalog.setContentTypeId(1);
+					catalog.setPages(content.getPages());
+					mBaseEntityList.add(catalog);
+					break;
+				case 2:
+					Manual manual = new Manual();
+					manual.setContentId(content.getContentId());
+					manual.setName(content.getName());
+					manual.setUrl(content.getUrl());
+					manual.setFileName(content.getFileName());
+					manual.setSize(content.getSize());
+					manual.setContentTypeId(2);
+					manual.setPages(content.getPages());
+					mBaseEntityList.add(manual);
+					break;
+				case 3:
+					Video video = new Video();
+					video.setContentId(content.getContentId());
+					video.setName(content.getName());
+					video.setUrl(content.getUrl());
+					video.setFileName(content.getFileName());
+					video.setSize(content.getSize());
+					video.setContentTypeId(3);
+					video.setDuration(content.getDuration());
+					mBaseEntityList.add(video);
+					break;
+			}
 
 		}
 	}
@@ -125,6 +178,25 @@ public class HistoryActivity extends Activity {
 			@Override
 			public void create(SwipeMenu menu) {
 				if (menu.getViewType() == CONTENT_TYPE_VIDEO) {
+					SwipeMenuItem download = new SwipeMenuItem(mContext);
+					download.setId(30000);
+					download.setBackground(new ColorDrawable(Color.rgb(0xF3, 0xF3, 0xF3)));
+					download.setWidth(Nexxoo.dp2px(mContext, isNormal ? 90 : 120));
+					download.setIcon(menu.getViewType() > 0?R.drawable.ic_list_trash:R
+							.drawable.ic_list_download);
+					download.setIsVisiable(true);
+					menu.addMenuItem(download);
+				}else{//PDF
+					SwipeMenuItem download = new SwipeMenuItem(mContext);
+					download.setId(30000);
+					download.setBackground(new ColorDrawable(Color.rgb(0xF3, 0xF3, 0xF3)));
+					download.setWidth(Nexxoo.dp2px(mContext, isNormal ? 90 : 120));
+					download.setIcon(R.drawable.ic_list_trash);
+					download.setIsVisiable(menu.getViewType() > 0);
+					menu.addMenuItem(download);
+				}
+
+				if (menu.getViewType() == CONTENT_TYPE_VIDEO) {
 					SwipeMenuItem play = new SwipeMenuItem(mContext);
 					play.setId(40000);
 					play.setBackground(new ColorDrawable(Color.rgb(0xE5, 0xF5, 0xFF)));
@@ -133,14 +205,6 @@ public class HistoryActivity extends Activity {
 					play.setIsVisiable(true);
 					menu.addMenuItem(play);
 				} else {
-					SwipeMenuItem download = new SwipeMenuItem(mContext);
-					download.setId(30000);
-					download.setBackground(new ColorDrawable(Color.rgb(0xF3, 0xF3, 0xF3)));
-					download.setWidth(Nexxoo.dp2px(mContext, isNormal ? 90 : 120));
-					download.setIcon(R.drawable.ic_list_trash);
-					download.setIsVisiable(menu.getViewType() > 0);
-					menu.addMenuItem(download);
-
 					SwipeMenuItem view = new SwipeMenuItem(mContext);
 					view.setId(40000);
 					view.setBackground(new ColorDrawable(Color.rgb(0xE5, 0xF5, 0xFF)));
@@ -169,6 +233,7 @@ public class HistoryActivity extends Activity {
 								File video = new File(mFileStorgeHelper.getDownloadAbsolutePath
 										(mBaseEntityList
 												.get(position).getFileName()));
+								if(video.exists())
 								video.delete();
 								/**
 								 * delete content from DB
@@ -183,6 +248,14 @@ public class HistoryActivity extends Activity {
 								ImageView image = (ImageView) imageLayout.getChildAt(0);
 								image.setImageResource(R.drawable.ic_list_download);
 							} else {// not downloaded
+								/**
+								 * add download content to local DB
+								 */
+								//should add wifi status check
+								Video video = (Video)mBaseEntityList.get(position);
+								ContentDBHelper db = new ContentDBHelper(mContext);
+								db.addContact(video);
+
 								DownloadAsyncTask task = new DownloadAsyncTask(mContext,
 										mBaseEntityList
 												.get(position).getUrl(), mBaseEntityList.get
@@ -198,15 +271,18 @@ public class HistoryActivity extends Activity {
 						case 50001://play button
 							isVideoDownloaded = mFileStorgeHelper.isContentDownloaded(mBaseEntityList
 									.get(position).getFileName());
-							String url = mBaseEntityList.get(position).getUrl();
-							url.replace("www", "nexxoo:wenexxoo4kiosk!@www");
-							Intent i = new Intent(mContext, VideoActivity.class);
-							i.putExtra(mContext.getString(R.string
-									.video_activity_intent_url_extra), url);
-							String name = mBaseEntityList.get(position).getFileName();
-							i.putExtra("filename", name);
-							i.putExtra("isVideoDownloaded", isVideoDownloaded);
-							mContext.startActivity(i);
+							String url1 = mBaseEntityList.get(position).getUrl();
+
+							Intent i1 = new Intent(mContext, VideoActivity.class);
+							if (url1 != null) {
+								url1.replace("www", "nexxoo:wenexxoo4kiosk!@www");
+								i1.putExtra(mContext.getString(R.string
+										.video_activity_intent_url_extra), url1);
+							}
+							String name1 = mBaseEntityList.get(position).getFileName();
+							i1.putExtra("filename", name1);
+							i1.putExtra("isVideoDownloaded", isVideoDownloaded);
+							mContext.startActivity(i1);
 							break;
 					}
 				} else {//PDF
@@ -231,7 +307,8 @@ public class HistoryActivity extends Activity {
 								//invisible trash icon
 								break;
 							case 50001:
-								String filename = mBaseEntityList.get(position).getFileName();
+								BaseEntity base = mBaseEntityList.get(position);
+								String filename = base.getFileName();
 								if (mFileStorgeHelper.isContentDownloaded(filename)) {
 									listview.closeAllMenu();
 									File file = new File(mFileStorgeHelper.getDownloadAbsolutePath(filename));
@@ -244,42 +321,13 @@ public class HistoryActivity extends Activity {
 								} else {
 									listview.closeAllMenu();
 									DownloadAsyncTask task1 = new DownloadAsyncTask(mContext,
-											mBaseEntityList
-													.get(position).getUrl(), mBaseEntityList.get
-											(position).getFileName());
+											base.getUrl(),
+											base.getFileName());
 									task1.execute();
 								}
 								break;
 						}
-					} else {// one button
-						LinearLayout image = (LinearLayout) parent.findViewById(new
-								Integer(50000));
-						image.setVisibility(View.VISIBLE);
-
-						String filename1 = mBaseEntityList.get(position).getFileName();
-						if (mFileStorgeHelper.isContentDownloaded(filename1)) {
-							File file = new File(mFileStorgeHelper.getDownloadAbsolutePath
-									(filename1));
-							Intent target = new Intent(Intent.ACTION_VIEW);
-							target.setDataAndType(Uri.fromFile(file), "application/pdf");
-							target.setFlags(Intent.FLAG_ACTIVITY_NO_HISTORY);
-							Intent i = Intent.createChooser(target, mContext
-									.getString(R.string.open_pdf_file));
-							startActivity(i);
-						} else {
-							if (listview.getChildAt(position) instanceof SwipeMenuLayout) {
-								SwipeMenuLayout menuLayout = (SwipeMenuLayout)
-										listview.getChildAt(position);
-								menuLayout.smoothCloseMenu();
-							}
-							DownloadAsyncTask task1 = new DownloadAsyncTask(mContext,
-									mBaseEntityList
-											.get(position).getUrl(), mBaseEntityList.get
-									(position).getFileName(), Global
-									.DOWNLOAD_TASK_TYPE_PDF);
-							task1.execute();
-						}
-					}
+					} //on history list all PDF are already be downloaded.
 				}
 				return false;
 			}
@@ -299,7 +347,14 @@ public class HistoryActivity extends Activity {
 		});
 	}
 
+	@Override
+	public void onBackPressed() {
+		super.onBackPressed();
+
+	}
+
 	private void prepareListData(JSONObject json) {
+		mBaseEntityList.clear();
 		try {
 			int count = json.getInt("count");
 			Manual manual = null;
@@ -358,7 +413,25 @@ public class HistoryActivity extends Activity {
 	public boolean onOptionsItemSelected(MenuItem menuItem) {
 		switch (menuItem.getItemId()) {
 			case android.R.id.home:
-				finish();
+				if (!Misc.isOnline(this)) {
+					new AlertDialog.Builder(this)
+							.setMessage(this.getResources().getString(R.string.no_wifi_message))
+							.setPositiveButton(android.R.string.yes, new DialogInterface.OnClickListener() {
+								public void onClick(DialogInterface dialog, int which) {
+								}
+							})
+							.setIcon(android.R.drawable.ic_dialog_alert)
+							.show();
+				}else{
+					if(fromMenu)
+						finish();
+					else{
+						Intent i = new Intent(HistoryActivity.this, MainActivity.class);
+						i.putExtra(getString(R.string.wifi_status), false);
+						startActivity(i);
+					}
+
+				}
 		}
 		return (super.onOptionsItemSelected(menuItem));
 	}
